pyproject.toml
[project]
name = "askaaron"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cachetools>=6.1.0",
    "fastapi[standard]>=0.116.1",
    "jinja2>=3.1.6",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "sqlmodel>=0.0.24",
    "uvicorn>=0.35.0",
]
